backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "ai-ppt-maker-backend"
version = "0.1.0"
description = "FastAPI backend for AI PPT Maker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "openai>=1.0.0,<2.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "pymupdf>=1.24.0,<2.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "python-pptx>=1.0.2,<2.0.0",
  "razorpay>=1.4.2,<2.0.0",
  "supabase>=2.7.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.0.0,<9.0.0",
  "ruff>=0.5.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
